NeuroPhish/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg
psycopg2-binary==2.9.9
greenlet
aiosqlite
redis==5.0.1

# ML and AI
torch==2.1.1
transformers==4.36.0
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.43.0
ipython
sentence-transformers==2.2.2
onnx==1.15.0
onnxruntime==1.16.3

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0

# Audio processing
openai-whisper==20231117
vosk==0.3.44
librosa==0.10.1

# HTTP clients and APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
beautifulsoup4
python-dotenv==1.0.0
celery==5.3.4
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0

# PDF generation
reportlab==4.0.7
weasyprint==60.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Email parsing
email-validator==2.1.0
# python-email-mime==0.1.3

# URL analysis
python-whois
tldextract==5.1.1
publicsuffix2==2.20191221

# File format support
python-magic==0.4.27
SpeechRecognition
pydub
